# target/classes/openapi/kraken-models-api-spec.yaml
openapi: 3.0.0

info:
  version: 1.0.0
  title: Kraken Orderbook Service
  description: This is the YAML For Kraken Orderbook Service

servers:
  - url: http://localhost:8080
    description: Local machine
tags:
  - name: "Orderbook"
    description: All Orderbook methods

paths:
  /orderbook/current:
    get:
      summary: Get the current order book snapshot.
      tags:
        - Orderbook
      parameters:
        - $ref: '#/components/parameters/currency-pair-in-query'
      responses:
        '200':
          description: Order book data retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderBook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    ErrorResponse:
      description: This is the response object in case of errors, compliant with RFC7807
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorData"
    ErrorData:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    OrderBook:
      type: object
      properties:
        bids:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        asks:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    Order:
      type: object
      properties:
        price:
          type: number
          format: double
          description: Price per unit.
        volume:
          type: number
          format: double
          description: Volume available at this price.
  parameters:
    currency-pair-in-query:
      in: query
      name: currencyPair
      required: true
      schema:
        type: string
        enum: [ BTCUSD, ETHUSD ]
        description: Currency pair to fetch the order book for.
  responses:
    BadRequest:
      description: "BAD REQUEST"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: "FORBIDDEN"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "NOT FOUND"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: "INTERNAL SERVER ERROR"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"